+++ sihl-user/src/sihl_user.ml
@@ -212,7 +212,9 @@ module Make (Repo : User_repo.Sig) : Sihl.Contract.User.Sig = struct
       Web.user_from_session (fun user_id -> find_opt ~user_id)
     ;;
 
+    let user_to_session = Web.user_to_session
     let user_from_token = Web.user_from_token (fun user_id -> find_opt ~user_id)
+    let user_to_token = Web.user_to_token
   end
 end